Use crate thread-scoped for BackgroundSession
The standard library dropped scoped() completely because of its safety
faults.  The crate thread-scoped provides the same functionality, but as
an unsafe function.  Use that, trickling the unsafe qualifier all the
way back to the spawn_mount() caller, so users are responsible for their
own safety guarantees here.  (i.e. Don't leak the JoinGuard!)

Fixes #50.