Remove ReplyDirectory::sized().
This function is public, which implies that clients can change the size
of the reply buffer after being called by FUSE. In reality, this doesn't
work; the size is dictated by FUSE, and changing it results in an error.

Though this is an API change, any clients broken by this were broken
already -- they had a bug if they assumed this function worked.