Use OsStr instead of Path to pass names to filesystem methods.
Names passed to fs methods are not paths, but filenames within the given parent inode, so the proper type is OsStr rather than Path. The only real path passed is the target of a symlink. Path was chosen a long time ago when there was no OsStr in Rust and we needed to pass filename-compatible strings.