Remove mountpoint reference from {,spawn_}mount()
The mount() and spawn_mount() calls define mountpoint as a &P but
P is already an AsRef<Path>.  This prevents passing the mountpoint
to these calls when it is already a reference without having to take
an additional reference to it.  For example:

fn foo(mountpoint: &Path) {
    fuse::mount(..., mountpoint, ...);
    ...
}

does not compile without passing mountpoint as &mountpoint to the
fuse::mount() call, but there is no reason that it should not.

To fix this, remove the extra reference reference qualifier.