Make low-level request API more self-contained
With this change all the struct members of `ll::request::op::*` are now
private and are accessed to accessors.  The usual sync trait API is
unchanged, but is now a simple wrapper around the `ll::Request`.

The eventual intention is for ll::Request to become an API in its own
right on top of which the traditional sync trait or future async APIs can
be built.  It contains no I/O functions and only operates on buffers
provided to it (the sans-io approach).  Dealing with the enum based Request
API provides additional flexibility - particularly WRT threading and async.
The user will eventually be able to pull a chunk of data from the FUSE
device and send it to a rayon threadpool, or an async executor, etc. to
process it.

Every operation corresponds to an opaque struct.  We can make additions
without breaking backwards compatibility.  In particular in the future I'd
like to make it so we can support multiple ABI versions concurrently - and
this API supports that.

There is no public API break intended.