ll/request.rs: Introduce newtypes for FUSE object refs
Unique request ids are `RequestId`, inode numbers are `INodeNo`, file
handles are `FileHandle` and lock owners are `LockOwner`.

This is intended in the future to make it harder to make mistakes, as you
can't get one `u64` mixed up with another, but also to improve
documentation and discoverablility.  If you want to find out which
functions are handling file handles you could now just grep for it.

I've not threaded this through the sync trait API as that would be an API
break.  We can do this later.