Add sans-io `Reply` type
This is similar to `ll::Request`, but for replies.  It's like what we see
in `src/reply.rs`, but it's immutable and does no I/O itself.  This makes
it suitable for returning from functions, rather than being passed in like
`crate::Reply`.  `Reply` is type-erased - there is a single `Reply` type
for all request types.  Choosing the correct type is for higher layers,
and it exposes no public constructors.

This is intended to help with having multiple different I/O backends, like
a sync one or a multithreaded one or a tokio one in the future.