Extend reftests to test invalid names and explicit directories
We have a few different semantics choices to make here about how to deal
with overlapping file and directory names. To make these explicit this
change refactors how we generate the test bucket and reference
filesystem layout: we first flatten the generated tree to generate the
S3 objects, and then re-build the tree while applying our expected
semantics to get the reference tree.