First pass at metrics support
We use the [`metrics` facade](https://docs.rs/metrics) to emit metrics
from both the connector and S3 client. The connector has a metrics
aggregator that currently just prints data to a `tracing` `info!`
message every 5 seconds. We do a little bit of work to automatically
instrument FUSE requests with their latency, but other than that,
everything is manual.

Closes #8.