Add support for custom endpoint URLs
This allows us to support dualstack and FIPS endpoints, which I've added
tests for. It should also allow us to support multi-region access points
and Transfer Acceleration, but I haven't written tests for those just
yet.

The overall idea follows the Python SDK and the AWS CLI. Customers can
specify an endpoint URL and an addressing mode (virtual host or
path-based). The SDKs go further and do endpoint discovery based on
configuration, so customers can specify an ARN and/or various endpoint
configurations like dualstack or FIPS, and the SDK automatically
computes the right endpoint. We don't have that yet because we'd like to
do it without replicating the SDKs.

This change was reasonably invasive because we have some new places to
thread errors through, but the important stuff is the new `Uri` that
exposes the CRT's URI support, and the new `Endpoint` struct that
encapsulates the rules for constructing endpoint URIs.

Signed-off-by: James Bornholt <bornholt@amazon.com>