Revert "Force direct IO for all files (#84)"
This change is causing reads to sometimes return uninitialized memory.
Our integration tests fail about 50% of the time with this change, and
never with this change reverted. I don't think it's a bug on our
side--the data we're writing to the FUSE device is correct--but I need
to poke at it further. In the meantime, let's just revert until we
understand it better.

This reverts commit c8296cb067f314b85d9b902d0724de04a82ca4ab.

Signed-off-by: James Bornholt <bornholt@amazon.com>