Fix URL encoding of request paths
We weren't correctly URL encoding paths before passing them to the CRT,
which expects them to already be encoded. This meant keys with weird
characters wouldn't work properly. The rules for encoding S3 keys aren't
explicit in the documentation, so this is my best attempt to match what
the Python SDK does (using `urllib.parse.quote`).

Signed-off-by: James Bornholt <bornholt@amazon.com>