Speed up some slow unit tests
We have a few slow unit tests that are pretty annoying. They're not slow
for any good reason -- the time is all spent moving large buffers around
rather than actually testing the logic. This change just makes all the
constants for these tests a bit smaller, so we're still testing the same
logic, just with fewer bytes to move around.

Signed-off-by: James Bornholt <bornholt@amazon.com>