Add initial support for sequential writes to new objects
This is a first pass at write support specifically for sequential writes
into new objects. It's currently guarded behind a new "put" feature
flag, so by default there will be no change and our file system will
continue to be mounted read-only.

This version is focused on getting the inode creation and locking story
in order. We need to make sure we don't allow a file to be opened for
writing more than once, and that we don't allow a file to be read while
being written. The inode locking changes in #113 allow us to do this by
holding on to the directory lock while creating the new object.

Currently, writes are just buffered in memory, and uploaded as a single
PutObject request when the file is closed. In future changes we'll make
this path support streaming and MPU for large objects and better upload
performance.

There are a couple of unclear places where we need to make semantics
decisions:
* Closing a file in Linux is asynchronous -- `close` failures can't be
  retried by applications. So applications have no way of knowing if a
  PUT fails. We probably need to explicitly document this, and also
  implement `fsync` so applications that care can flush (and find out
  about the failure) before closing.
* The right error numbers for some of the failures are unclear. Linux
  doesn't really anticipate a world where a file can't be read but can
  be written, so we try to choose reasonable errnos (EPERM) but they're
  not perfect. Again this is something we should document.
* If an application seeks forwards while writing, should we zero-fill
  those skipped offsets and then reject future writes to them? Right now
  this will just fail, but perhaps we can do better.

Signed-off-by: James Bornholt <bornholt@amazon.com>