Remove O_DIRECT from basic_read_test
This test is failing sporadically in CI with corrupt data. I believe the
cause is the same as #84. Per the man page for `open(2)`, `O_DIRECT` is
not safe to use concurrently with `fork` syscalls:

> O_DIRECT I/Os should never be run concurrently with the fork(2)
> system call, if the memory buffer is a private mapping (i.e., any
> mapping created with the mmap(2) MAP_PRIVATE flag; this includes
> memory allocated on the heap and statically allocated buffers).
> Any such I/Os, whether submitted via an asynchronous I/O
> interface or from another thread in the process, should be
> completed before fork(2) is called.  Failure to do so can result
> in data corruption and undefined behavior in parent and child
> processes.

Our test runner does a `fork` every time we mount a FUSE file system,
since we don't run our tests as root, and so we need to execute the
`fusermount3` utility to do the mount. Since tests run concurrently,
that means we will sometimes do one of these forks at the same time as
this test is doing an `O_DIRECT` read.

Signed-off-by: James Bornholt <bornholt@cs.utexas.edu>