Refactor inode interface to prepare for writes (#113)
* Refactor inode interface to prepare for writes

This commit does some refactoring of our inode story to prepare for the
next PR that will introduce write support. There's no functional change
in this commit, just refactoring and the consequences of that.

The big idea is that an `Inode` is now a pointer to the actual inode
data, so we can cheaply copy them around and hold onto references to
them without worrying about lifetimes. Before, all inodes were stored in
a big locked hashmap, so if we wanted to hold onto an inode reference we
had to instead remember its inode number, and then do a re-lookup in the
hashmap any time we needed it. We were relying on this behavior to
serialize some operations like `update_or_insert`, but it became very
clumsy as I was adding writes.

Inodes now have a locked `InodeState`, which allows us to serialize updates
to a particular inode in isolation from the big locked hashmap. This
makes `update_or_insert` much clearer. It also makes the future checks
for whether an inode is writable or not simpler to think through.

The client of `inode.rs` now gets access to the `Inode` struct rather
than just seeing everything as inode numbers. But it only gets access
through some getters rather than direct access to the fields.

One regression in this change is that we now make a complete copy of the
key for each inode. Before, each inode only stored its own component of
the path, and we reconstructed full keys on the fly by walking up the
directory hierarchy. The new version is more memory inefficient, but
makes the locking story simpler. It shouldn't be too hard to turn this
into some kind of string interning thing in the future to recover the
memory overhead.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Make `InodeInner` a named field

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>