Consistently use UTF-8 for S3 keys
S3 object keys are UTF-8 strings, but Unix path names are (mostly)
arbitrary byte strings, so we'll always have to do some conversions
between the two. Right now that conversion is ad hoc and done in several
places. This commit standardizes all our `inode` code to do UTF-8 checks
as soon as an `OsStr` is received, and then work entirely in terms of
`str`s.

Signed-off-by: James Bornholt <bornholt@amazon.com>