Make `Filesystem` trait use interior mutability
Currently the Filesystem trait forces only a single request dispatch
loop, with the idea that a filesystem that wants to be concurrent can
just spawn its own threads and dispatch operations on them itself. This
works, but it's bad for performance, because it forces every request to
bounce across threads.

Instead, let's make the Filesystem trait use interior mutability. This
allows us to spawn multiple dispatch threads that each block on the FUSE
device, read a request, and immediately dispatch it without the bounce.

Unfortunately there's not a good way for a trait to be variant in
mutability, so this is a breaking API change. We'll need to think more
carefully about how to upstream this change.

Signed-off-by: James Bornholt <bornholt@amazon.com>