Switch to cleaned up non-async fuser fork
Our current fuser fork has a messy commit history and does two big
changes at once: (1) concurrent filesystems and (2) async filesystems.
We don't actually need the async version right now -- it was incomplete
and we were just running it on blocking threads anyway.

This change redirects the fuser fork to our `fuser/fork` branch, which
is a cleaned up history of the fork but without async support. Instead,
we push the async support down into our implementation via `block_on`.
This performs the same, but with a smaller change to fuser.

In the future we'll still want to async-ify fuser so the entire stack
can run on top of (the CRT's?) async runtime. But there's more work
involved in correctly async-ifying fuser, like doing async IO on the
FUSE device, so we can defer that for now.

This change temporarily breaks the `--thread-count` command-line
argument and so will tank concurrent performance tests. I'll fix that in
a followup as I don't want this diff to get any bigger, and will fix
unmount support (#93) at the same time.

Signed-off-by: James Bornholt <bornholt@amazon.com>