Avoid copying reply buffer in `ReplyData`
The reply already doesn't take ownership of the buffer, and it just ends
up being written synchronously to the FUSE device, so there's no need to
make a copy of it. Instead, the `Response` can just hold onto the slice
and use it directly when building the IoVec for writing the response.