Split integration tests into a new CI workflow
Our integration tests are tricky to run because they require S3
credentials, which GitHub (quite rightly) makes difficult to access
through a pull request from a fork repo. Even if we turn on GitHub's
feature for requiring manual approval for a PR to run workflows, it
still won't have access to the permissions it needs to get S3
credentials, because the `configure-aws-credentials` action needs
`id_token: write` permissions, which fork PRs can never have.

This change makes our integration tests run using the
`pull_request_target` event. This event runs in the context of the base
repo rather than the PR, and so has access to the right permissions.
That's dangerous, of course, so we want to manually approve runs of this
workflow. But `pull_request_target` workflows are _automatically
approved_, so the normal "require manual approval for a PR to run
workflows" no longer works. Instead, we rig something up using
deployment environments, which can be configured to require manual
approval before deployments. The integration tests ask to be "deployed"
to an environment with approval configured, and so won't run until
someone with permission to approve those deployments clicks OK. Same
effect as the normal PR approval mechanism, just a little more
jury-rigged and with some confusing verbiage around "deployment".

Pushes to `main` also run the integration tests, but they ask for a null
deployment, and so run as normal workflows. This is safe because pushes
to `main` already require PR approvals.

Workflows on `pull_request_target` have a different security boundary
[1], and in particular have two permissions normal `pull_request`
workflows don't have:
1. Their GITHUB_TOKEN has write access to the repository. We
   explicitly downgrade the token to have only read access using the
   `permissions` field in integration test workflows.
2. They run in the context of the base repository, and so have access to
   secrets. We don't use any secrets. We also explicitly prevent the
   integration tests from storing cached artifacts if running from a PR,
   which could poison other users of the cache.

Since they run in the base context, `pull_request_target` workflows will
always run with the workflow files from the base repo, ignoring any
changes in the pull request. That means a PR can't change the workflow
to undo these protections.

Fixes #143.

[1]: https://securitylab.github.com/research/github-actions-preventing-pwn-requests/

Signed-off-by: James Bornholt <bornholt@amazon.com>