Fix time dependency in mountpoint-s3-client
Just a tiny thing: this crate's use of `time` requires some features.
This isn't a problem for us because mountpoint-s3 pulls those features
in, so compilation still works, but it's a problem if someone wants to
take a direct dependency on mountpoint-s3-client. Also, there's no need
for the direct dependency on the -sys crate here.

Signed-off-by: James Bornholt <bornholt@amazon.com>