Make opened files visible in readdir (#189)
* Make opened files visible in readdir

Currently, new files that are opened for write will not be visible in
the file system until they are closed and their contents are uploaded to
S3, but we want them to be visible as soon as we open the files.

To do this, we will store inodes of the opened files in its parent state
and remove them once we finish writing. Then, on readdir, we can append
these inodes to the results we get from LIST API to make the opened files
visible in the file system.

Another problem is results ordering. Previously, we're using double-ended
queue to store the sorted LIST results but it doesn't work well with newly
created files since we don't know their exact positions in paginated
results. That's why it's also replaced by priority queue in this commit.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Address PR comments

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Use two VecDeques instead of BinaryHeap for results ordering

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Wrap the write state transition methods in WriteHandle

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Use random file names for readdir tests

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Improve testing

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>