Make opened file visible in lookup
We have implemented this in readdir but missed it in lookup. We want
the new files to be visible there as well. There might be two options
here, 1/ We can lookup locally first and return immediately when the
item we want is an uncommitted inode and we don't have to call any S3
APIs since it should not exist in S3 yet, or 2/ we can lookup remotely
first then fall back to local. In this commit, we go with the second
option because there is a possibility that someone could make a change
remotely like creating a directory marker that shadow the uncommitted
file.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>