Add benchmark for readdir (#213)
* Add benchmark for readdir

In this commit, we add a new benchmark script for readdir operation. The
benchmark itself is pretty simple, we're just doing ls against pre-generated
directories with different size. However, comparing the results is a bit
tricky because we're now measuring latency instead of throughput which
is what we do in the existing benchmark.

Our current CI workflows don't support this scenario, so we have to
introduce another workflow job just for this kind of benchmark. This new
job will run the latency benchmark script, compare the new result with
the previous one, and store them at dev/latency_bench path.

Max number of data points for the chart is also added in this commit to
keep the chart clean.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Refactor bench script

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Update benchmarking doc and add some fio files

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>