Add time to first byte benchmark
In this commit, we add a new type of benchmark to measure the latency
aspect of the read operation.

We do this by running fio workloads that send one byte read request with
one byte block size against existing files. The time it takes to complete
this operation will be measured and recorded as time to first byte read.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>