Implement checksum validation for prefetcher
We want to ensure data integrity while object data is within Mountpoint's
custody. So, we will calculate CRC checksum for the data as soon as we
receive it from S3 and validate them every time we move it around our
part queue, and before returning them to FUSE.

The validation would definitely impact Mountpoint's performance. But we
can improve it a bit if we assume that most of the sequential read would
have the same read size. In that case, we can pre-split the parts in part
queues to a preferred size and avoid splitting them during read.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>