Basic property-based tests for writes (#229)
* Basic property-based tests for writes

This is the first step in property-based testing for the write path.
Right now, it tests scenarios where the entire write sequence happens at
once: mknod, open, write once, release. It generates a vector of such
operations, with random paths and random file contents/sizes. After each
operations, the test checks that the file system has the expected
contents.

I'm stopping this PR here because it's enough to get the testing off the
ground, but doesn't yet cover a lot of the interesting cases we'd like to
check. The focus here is getting the infrastructure in place to start
adding more complex operations to the `Op` structure in future PRs.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Don't do manual path manipulation + other PR comments

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Fix root directory

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>