Implement expiry times for `lookup` and `getattr` results
Today we don't quite have our management of TTLs correct, especially in
`getattr`, which never refreshes attributes. We also want to start using
TTLs correctly to improve performance in common cases like listing a
directory with `ls -l`, which issues a `stat` for every directory
entry.

The overall goal of this change is to correctly track TTLs within the
inode code and return them to FUSE. We configure two types of TTLs, one
for files and one for directories, because directory attributes are
queried far more often (for permissions). We make this work by adding a
`validity()` method to `LookedUp`, which FUSE operations query when they
need to indicate the validity of an entry they're returning. Then, in
`getattr`, we check if the attributes are stale before returning them.

With this change, running `ls -l` on a nested directory with 1000 entries
goes from taking 165 seconds to < 1 second.

I have three followup changes planned here, but this diff was already big
enough:
1. Document the new consistency behavior `in doc/SEMANTICS.md`.
2. Write tests for the TTL behavior. This is too hard to do right now
   because the tests are very sensitive to timing. I think we'll want a
   way to mock the clock so we can control it.
3. I think we can just totally remove the `stat` field from remote
   inodes. The idea is that FUSE only ever calls `lookup` or `getattr`
   if the cache has already expired, so keeping a cached copy of the stat
   around is useless (it'll always be expired when queried). Removing this
   field should make the inode code a bunch simpler since there's no longer
   a need to track/update stats for remote files.

Signed-off-by: James Bornholt <bornholt@amazon.com>