Refactor ReaddirHandle to lazily create inodes  (#238)
* Move ReaddirHandle into its own file

`inode.rs` is getting too long to be readable.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Refactor `ReaddirHandle` to lazily create inodes

The motivation here is that `readdir` calls from FUSE generally only
retrieve a small number of entries, but ListObjectsV2 calls to S3
retrieve many entries. In the current code, we instantiate inodes for
those entries as soon as the ListObjectsV2 call completes, but because
of the small batch size of `readdir`, most of these inodes will not be
returned to FUSE until some time later. Because the TTL of an inode's
attributes is chosen at inode creation time, it will have expired by
this point, and so the kernel will immediately re-`lookup` these entries
when they're returned to `readdir`. That makes `readdir` do O(n)
`lookup` calls back to S3, defeating the purpose of ListObjectsV2.

To solve this problem, this change refactors the `ReaddirHandle`
implementation to defer creating inodes until an entry is actually
returned to `readdir`. The big idea is that we now just create a new
[ReaddirEntry] struct for each entry returned by ListObjectsV2, and all
the `readdir` logic works in terms of those things rather than directly
on inodes. The conversion from [ReaddirEntry] to an actual inode then
happens right before returning the entry to `readdir`.

I took this opportunity to refactor this code (hopefully for the
better!) since this change was adding yet another layer of complexity
into the logic for iterating directory entries. I also tried to capture
my understanding of this code in a new module comment. There is still a
bug here around page sizes (#190) that I think this comment and refactor
will make it easier to fix, but let's do that separately to avoid making
this change bigger.

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>