Add incremental writes to reftests (#249)
* Add incremental writes to reftests

In #229 we added initial support for writes to the reference model
tests. So far, those tests have only covered writing to a file as a
single atomic operation (create, open, write, close), but we'd also
like to test interleavings of those individual operations, so that we
can see, for example, multiple writers try to touch the same file.

This change adds the create/open/write/close steps as separate
operations that the reftest can choose to perform in any order. We
retain the original atomic operation, but it's now just a "shortcut"
that does all four steps together.

The important thing we do here is introduce a couple of "index" types
that `proptest` can generate. These indexes allow the generated test to
refer to inflight writes, existing directories, etc all without actually
needing to understand paths. The indexes are strongly typed even though
they're just wrappers around `usize` so that we don't mix them up.

This change is hopefully also a template for how to add additional
operations in the future: we need to add new variants to `Op`, implement
a `perform_` method for them, and think about how to track any inflight
state so that future operations can refer to it.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Address PR feedback

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>