Implement checksummed bytes
This change introduces ChecksummedBytes struct which is a buffer containing
both the bytes and its checksum. There will be follow-up changes that make
use of this struct, as we planned to implement some integrity check along
the read and write path, the ChecksummedBytes will make those changes
simpler and more efficient.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>