Allow custom credentials providers for S3CrtClient (#252)
* Allow custom credentials providers for S3CrtClient

This slightly generalizes the work we've done to expose profiles and
no-sign-request by allowing a custom credentials provider. Right now,
only the CRT providers that we've explicitly bound can be used here, but
eventually we'll bind the `aws_credentials_provider_delegate` version
that can invoke an arbitrary callback to fetch credentials. For now,
added a static provider that can be manually configured with credentials
for customers using the client as a library with credentials vended
from elsewhere.

I also took this chance to write a test for the profile credentials
provider. It's a little annoying to do because CLI profiles are global
state, so we do it by forking the test into a new process where it's
safe to futz with the environment variables we need. Along the way, I
realized that the CRT auth provider prefers the AWS_PROFILE environment
variable over `--profile`, which is the wrong behavior, so I fixed that
by unsetting the variable in main.rs.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* PR feedback

Signed-off-by: James Bornholt <bornholt@amazon.com>

* Update mountpoint-s3-crt/src/auth/credentials.rs

Co-authored-by: Daniel Carl Jones <danny@danielcarl.info>
Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>
Co-authored-by: Daniel Carl Jones <danny@danielcarl.info>