Improve file handles locking (#253)
* Improve file handles locking

Currently, we're holding the lock on the file handles hash table for the
whole read and write operations. It blocks other operations from interacting
with the hash table, especially at high concurrency. Instead, we can hold
the read lock for only a short period and use the file handles via atomic
reference.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Add error log

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>