Implement checksum validation for the prefetcher
We want to ensure data integrity while it is within Mountpoint's custody
but S3 API doesn't support ranged-get checksum at the moment.

As a temporary measure, we will calculate CRC checksum for the data as
soon as we receive it from S3 and do the validation every time we pass it
around our part queue, and right before returning them to FUSE. To do
this, we will use ChecksummedBytes from previous change instead of Bytes
in the prefetcher.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>