Prepare to publish client and CRT crates
This change adds READMEs and CHANGELOGs for the client, crt, and crt-sys
crates, and bumps their version numbers just for clarity. It changes the
path dependencies to also include version numbers, so that local builds
of Mountpoint will use the path dependency but published releases will
use the published dependencies.

Signed-off-by: James Bornholt <bornholt@amazon.com>