Avoid CRT auto-ranged-get for small requests
The CRT's auto-ranged-get will always do a HeadObject request when a
range is specified. For small requests, that's a fairly large latency
hit. Let's avoid using the auto-ranged-get for requests small enough
that they would be sent as a single part anyway.

It would probably be nicer to have a way to turn off this behavior in
the CRT, but this is a reasonable change for now. We'll have to be
careful if we ever start relying on any extra behavior of the
auto-ranged-get machinery, like checksum validation, but we're not
currently planning anything like that.

Signed-off-by: James Bornholt <bornholt@amazon.com>