Force compiler optimizations for aws-checksums (#284)
The ARM implementations in aws-checksums are written in C, and so
compiler optimizations are essential to give them reasonable
performance. We've recently started validating a lot of checksums in our
tests (#263), and that's made the ARM CI much slower (> an hour). This
should get that back under control. x86 isn't affected because the
aws-checksums implementations are mostly hand-written assembly.

Signed-off-by: James Bornholt <bornholt@amazon.com>