Release new crate versions
The main goal here is to get #298 out the door, but it's a good point to
get the last month of updates out too.

The async streaming work changed the `put_object` interface, so this
release is a breaking change for `mountpoint-s3-client`.

We also missed writing a changelog for v0.2.2 of `mountpoint-s3-client`,
so I'm writing it here.

Signed-off-by: James Bornholt <bornholt@amazon.com>