Fix va_list FFI safety on aarch64
When building on aarch64 we get a warning about FFI safety for va_list.
It turns out bindgen can't figure out the definition for va_list on that
architecture and so just emits a `[u64; 4]`, which has the right size,
but raw arrays aren't FFI-safe. So instead, let's emit a `struct
va_list` that matches the aarch64 definitition. The actual fields don't
matter since we never actually try to look inside a `va_list`; we only
care that the bytes roundtrip correctly back to C.

Signed-off-by: James Bornholt <bornholt@amazon.com>