reftests: always materialize reference model from bucket contents (#311)
* reftests: always materialize reference model from bucket contents

Today, the reference model for the reftests is just a tree structure.
When we make changes to the expected state of the file system, we update
this tree structure in place.

This needs to change to prepare us to extend the reftests with mutations
to the *bucket*, not just the file system (i.e., concurrent mutations).
When one of these mutations happens, we need to be able to compute what
we expect the state of the file system to be after the change. For
example, if my bucket contains keys `a` and `a/b`, and then I delete
`a/b` from the bucket, what's the expected state of the file system? The
current implementation asks us to compute this state by making in-place
modifications to the file system tree, which is error prone -- we'd
essentially be re-implementing the same inode management logic we use in
the actual file system.

To avoid this complexity, this change instead "materializes" the file
system structure from the bucket structure at every step of the reftest.
The reference model is now just an S3 bucket plus lists of local files
and directories. In the example above, we can simulate deleting an
object by just removing it from both the mock bucket (the one Mountpoint
is accessing through the mock client) and the reference bucket (the one
introduced in this change). Then, when we rematerialize the reference
file system, we'll have the new directory structure, without having to
manually compute the changes to make to that tree.

This is nice because it means the only "trusted" code in the reference
model is `build_reference`, which specifies how to convert *any* bucket
into a filesystem. One potential downside is that rematerializing the
filesystem after every operation might be expensive, but in my testing
the overhead hasn't seemed problematic.

This commit is not expected to change any semantics of the reftests, and
doesn't add any new operations. Those will come later, building on this
refactoring.

Signed-off-by: James Bornholt <bornholt@amazon.com>

* PR feedback

Signed-off-by: James Bornholt <bornholt@amazon.com>

---------

Signed-off-by: James Bornholt <bornholt@amazon.com>