Add unlink to reference model tests
This is pretty annoying because of some weird edge cases around implicit
directories being removed while local children are present. I think in
the long term we want to fix this, but it's a similar problem to what we
saw in #359 -- we need `readdir` to clean up removed directories
properly. So for now, I've changed the reference model to match our
current semantics, which is that if an ancestor of a local
file/directory is removed, that file/directory will no longer be visible
through the filesystem. Of course, once that file/directory becomes
remote it will become visible, and the model still captures that.

Signed-off-by: James Bornholt <bornholt@amazon.com>