Implement fsync and handle write errors (#313)
Implement `fsync` to allow users to complete a put request and receive confirmation that it succeeded or failed. If a file handle is released without a call to `fsync`, `release` will still complete the upload as before. 

Wrap the `UploadRequest` in the file handle in a new `UploadState` enum, in order to detect:
* on `release`, whether the request had been already completed by an `fsync` call,
* `write` is invoked after an `fsync`,
* `write` (or `fsync`) is invoked after a previous call failed.

Also adds support for put failures to FailureClient.

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>