Optimize part size for checksummed read (#315)
* Optimize part size for checksummed read

The prefetcher stores data received from each input stream as a part in
the part queue structure. Usually, the part size is pretty big (8 MB or
more) and the checksum validation always has to be done against an entire
part even if we only read a small portion of that part.

This makes checksummed read much slower than non-checksummed read. We could
make it more efficient by making the part smaller or ideally align the part
size to the read size so that we don't have to compute the checksum on
unnecessary bytes.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

* Address PR comments

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>