Enforce maximum upload size
Uploads can fail if they require more than 10,000 parts. Right now we
only find out about that failure asynchronously when the CRT actually
constructs the 10,0001th part. This is bad because unlike other upload
errors, this one is a deterministic failure that we should be able to
report up front.

One trick here is that the kernel appears to retry failed writes from
page cache when returning EFBIG, so we need to remember the reason the
write failed.

Signed-off-by: James Bornholt <bornholt@amazon.com>